pyproject.toml
[tool.poetry]
name = "coinhuntbot"
version = "0.1.0"
description = ""
authors = ["Lopkop <lopkophacked@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.14.3"
python-dotenv = "^0.19.0"
beautifulsoup4 = "^4.9.3"
selenium = "^3.141.0"
asyncpg = "^0.24.0"
emoji = "^1.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
